requirements.txt
flask==3.0.3
gunicorn==22.0.0
PyYAML==6.0.2
pytest==8.4.1